refactor(gym): 제휴 헬스장 등록 기능 전반 구조 정비 및 요청/응답 리팩터링
- Request DTO를 기능별로 분리 (GymInfoRequest, UpdateGymRequest, RegisterGymRequest, GymRequest)
- 기존 Response 제거 및 필요 없는 구조 정리
- GymMapper 메서드들 수정 및 추가
- PartnerGymService에 등록/수정 관련 메서드 추가 (registerPartnerGym, updatePartnerGym, createPartnerGym)
- Controller에서 이미지 업로드를 multipart 형식으로 받도록 수정
