+++ src/main/java/org/helloworld/gymmate/common/config/SecurityConfig.java
@@ -71,7 +71,7 @@ public class SecurityConfig {
 				"/v3/**", "/swagger-ui/**", "/api/logistics",
 				"h2-console/**", "/error",
 				// 이하는 개발환경에서만 허용해야 함
-				"/gym/**", "/ptProduct/**", "/crawl/**"
+				 "/ptProduct/**", "/crawl/**"
 			)
 			//.requestMatchers(HttpMethod.GET, "/similar")
 			//.requestMatchers(HttpMethod.GET, "/reviews/{reviewId}/comments")
