--- src/main/java/org/helloworld/gymmate/domain/gym/gym/dto/request/PartnerGymRequest.java
@@ -1,4 +1,0 @@
-package org.helloworld.gymmate.domain.gym.gym.dto.request;
-
-public record PartnerGymRequest(Long gymId) {
-}
