--- src/main/java/org/helloworld/gymmate/domain/gym/gym/dto/response/PartnerGymResponse.java
@@ -1,4 +1,0 @@
-package org.helloworld.gymmate.domain.gym.gym.dto.response;
-
-public record PartnerGymResponse(Long partnerGymId) {
-}
