+++ src/main/java/org/helloworld/gymmate/domain/gym/gymInfo/dto/request/GymInfoRequest.java
@@ -1,0 +1,7 @@
+package org.helloworld.gymmate.domain.gym.gymInfo.dto.request;
+
+public record GymInfoRequest(
+	GymRequest gymRequest
+	//편의시설, 운동기구 dto 추가
+) {
+}
