+++ src/main/java/org/helloworld/gymmate/domain/gym/gymInfo/dto/request/GymRequest.java
@@ -1,10 +1,8 @@
-package org.helloworld.gymmate.domain.gym.gym.dto.request;
-
-import java.util.List;
+package org.helloworld.gymmate.domain.gym.gymInfo.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record GymUpdateRequest(
+public record GymRequest(
 	@NotBlank(message = "헬스장 이름을 입력해주세요.")
 	String gymName,
 
@@ -20,15 +18,7 @@ public record GymUpdateRequest(
 	@NotBlank(message = "주소를 입력해주세요.")
 	String address,
 
-	@NotBlank(message = "x 좌표를 입력해주세요.")
-	Double xField,
-
-	@NotBlank(message = "y 좌표를 입력해주세요.")
-	Double yField,
-
 	@NotBlank(message = "헬스장 소개를 입력해주세요.")
-	String intro,
-
-	List<Long> deleteImageIds
+	String intro
 ) {
 }
