+++ src/main/java/org/helloworld/gymmate/domain/gym/gymInfo/dto/request/RegisterGymRequest.java
@@ -1,0 +1,10 @@
+package org.helloworld.gymmate.domain.gym.gymInfo.dto.request;
+
+import java.util.List;
+
+public record RegisterGymRequest(
+	Long gymId,
+	GymInfoRequest gymInfoRequest,
+	List<String> imageUrl
+) {
+}
