+++ src/main/java/org/helloworld/gymmate/domain/gym/gymInfo/dto/request/UpdateGymRequest.java
@@ -1,0 +1,9 @@
+package org.helloworld.gymmate.domain.gym.gymInfo.dto.request;
+
+import java.util.List;
+
+public record UpdateGymRequest(
+	GymInfoRequest gymInfoRequest,
+	List<Long> deleteImageIds
+) {
+}
