+++ src/main/java/org/helloworld/gymmate/domain/gym/gymInfo/entity/PartnerGym.java
@@ -1,7 +1,6 @@
-package org.helloworld.gymmate.domain.gym.gym.entity;
+package org.helloworld.gymmate.domain.gym.gymInfo.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -15,27 +14,23 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-@Table(name = "gymImage")
-public class GymImage {
+@Table(name = "Partner_gym")
+public class PartnerGym {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Long partnerGymId;
 
-	private String url; // 이미지 경로
+	@Column(name = "owner_id", nullable = false)
+	private Long ownerId; //헬스장 주인id
 
-	//제한된 setter 사용
-	@Setter
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "gym_id")
-	@JsonIgnore
+	@JoinColumn(name = "gym_id", nullable = false)
 	private Gym gym;
-
 }
