+++ src/main/java/org/helloworld/gymmate/domain/gym/gymInfo/repository/PartnerGymRepository.java
@@ -1,6 +1,6 @@
-package org.helloworld.gymmate.domain.gym.gym.repository;
+package org.helloworld.gymmate.domain.gym.gymInfo.repository;
 
-import org.helloworld.gymmate.domain.gym.gym.entity.PartnerGym;
+import org.helloworld.gymmate.domain.gym.gymInfo.entity.PartnerGym;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PartnerGymRepository extends JpaRepository<PartnerGym, Long> {
